Combine code related to imagery layers into a single file
